Use shared PRNG for seeded Random
Move xoshiro128** into the internal random module and reuse it from Random.withSeed and Arbitrary. This removes the duplicate seeded generator while preserving Arbitrary replay and isolated per-attempt streams.

Compared with ISAAC, the local spike made seeded construction about 36.7x faster and direct draws about 24.7% faster. The measured gzip fixtures became 0.29 KB smaller for Random and 0.78 KB smaller when Random and Arbitrary were combined.

Document Crypto as the secure alternative because the shared generator is deterministic but not cryptographically secure. Seeded Random sequences intentionally change.
